refactor for more logical structure
