+++ packages/sdk/chainDeploy/helpers/index.ts
@@ -1,4 +1,9 @@
 export { ChainDeployConfig, ChainlinkFeedBaseCurrency } from "./types";
 export { deployIRMs } from "./irms";
-export { deployUniswapOracle } from "./uniswap";
-export { deployChainlinkOracle } from "./chainlink";
+export { deployUniswapOracle } from "./oracles/uniswap";
+export { deployUniswapLpOracle } from "./oracles/uniswapLp";
+export { deployCurveLpOracle } from "./oracles/curveLp";
+export { deployABNBcOracle } from "./oracles/aBNBcOracle";
+export { deployChainlinkOracle } from "./oracles/chainlink";
+export { deployDiaOracle } from "./oracles/dia";
+export { deployGelatoGUniPriceOracle } from "./oracles/gelato";
