+++ packages/sdk/chainDeploy/helpers/oracles/dia.ts
@@ -1,8 +1,7 @@
 import { providers } from "ethers";
 
-import { DiaPriceOracle } from "../../lib/contracts/typechain/DiaPriceOracle.sol";
-
-import { DiaDeployFnParams } from "./types";
+import { DiaPriceOracle } from "../../../lib/contracts/typechain/DiaPriceOracle.sol";
+import { DiaDeployFnParams } from "../types";
 
 export const deployDiaOracle = async ({
   ethers,
