+++ packages/sdk/chainDeploy/helpers/oracles/uniswapLp.ts
@@ -1,7 +1,7 @@
 import { constants } from "ethers";
 
-import { MasterPriceOracle } from "../../lib/contracts/typechain/MasterPriceOracle";
-import { UniswapDeployFnParams } from "../helpers/types";
+import { MasterPriceOracle } from "../../../lib/contracts/typechain/MasterPriceOracle";
+import { UniswapDeployFnParams } from "../../helpers/types";
 
 export const deployUniswapLpOracle = async ({
   ethers,
